prov/verbs: Move MSG, RMA, Atomics to one file
