+++ prov/verbs/Makefile.include
@@ -2,7 +2,6 @@ if HAVE_VERBS
 _verbs_files =							\
 	prov/verbs/src/fi_verbs.h				\
 	prov/verbs/src/fi_verbs.c				\
-	prov/verbs/src/verbs_atomic.c				\
 	prov/verbs/src/verbs_cm.c				\
 	prov/verbs/src/verbs_cq.c				\
 	prov/verbs/src/verbs_srq.c				\
@@ -10,9 +9,7 @@ _verbs_files =							\
 	prov/verbs/src/verbs_mr.c				\
 	prov/verbs/src/verbs_eq.c				\
 	prov/verbs/src/verbs_info.c				\
-	prov/verbs/src/verbs_msg.c				\
 	prov/verbs/src/verbs_msg_ep.c				\
-	prov/verbs/src/verbs_rma.c				\
 	prov/verbs/src/ep_rdm/verbs_av_ep_rdm.c			\
 	prov/verbs/src/ep_rdm/verbs_cq_ep_rdm.c			\
 	prov/verbs/src/ep_rdm/verbs_ep_rdm.c			\
