+++ prov/verbs/src/verbs_msg_ep.c
@@ -523,3 +523,813 @@ err1:
 	return ret;
 }
 
+static ssize_t
+fi_ibv_msg_ep_recvmsg(struct fid_ep *ep, const struct fi_msg *msg, uint64_t flags)
+{
+	struct fi_ibv_msg_ep *_ep;
+	struct fi_ibv_wre *wre;
+	struct ibv_sge *sge = NULL;
+	size_t i;
+
+	_ep = container_of(ep, struct fi_ibv_msg_ep, ep_fid);
+	assert(_ep->rcq);
+
+	fastlock_acquire(&_ep->wre_lock);
+	wre = util_buf_alloc(_ep->wre_pool);
+	if (OFI_UNLIKELY(!wre)) {
+		fastlock_release(&_ep->wre_lock);
+		return -FI_EAGAIN;
+	}
+	memset(wre, 0, sizeof(*wre));
+	dlist_insert_tail(&wre->entry, &_ep->wre_list);
+	fastlock_release(&_ep->wre_lock);
+
+	wre->ep = _ep;
+	wre->context = msg->context;
+	wre->wr.type = IBV_RECV_WR;
+	wre->wr.rwr.wr_id = (uintptr_t)wre;
+	wre->wr.rwr.next = NULL;
+
+	if (msg->iov_count) {
+		sge = alloca(sizeof(*sge) * msg->iov_count);
+		for (i = 0; i < msg->iov_count; i++) {
+			sge[i].addr = (uintptr_t)msg->msg_iov[i].iov_base;
+			sge[i].length = (uint32_t)msg->msg_iov[i].iov_len;
+			sge[i].lkey = (uint32_t)(uintptr_t)(msg->desc[i]);
+		}
+	}
+	wre->wr.rwr.sg_list = sge;
+	wre->wr.rwr.num_sge = msg->iov_count;
+
+	return FI_IBV_INVOKE_POST(recv, recv, _ep->id->qp, &wre->wr.rwr,
+				  FI_IBV_RELEASE_WRE(_ep, wre));
+}
+
+static ssize_t
+fi_ibv_msg_ep_recv(struct fid_ep *ep, void *buf, size_t len,
+		void *desc, fi_addr_t src_addr, void *context)
+{
+	struct iovec iov;
+	struct fi_msg msg;
+
+	iov.iov_base = buf;
+	iov.iov_len = len;
+
+	msg.msg_iov = &iov;
+	msg.desc = &desc;
+	msg.iov_count = 1;
+	msg.addr = src_addr;
+	msg.context = context;
+
+	return fi_ibv_msg_ep_recvmsg(ep, &msg, 0);
+}
+
+static ssize_t
+fi_ibv_msg_ep_recvv(struct fid_ep *ep, const struct iovec *iov, void **desc,
+                 size_t count, fi_addr_t src_addr, void *context)
+{
+	struct fi_msg msg;
+
+	msg.msg_iov = iov;
+	msg.desc = desc;
+	msg.iov_count = count;
+	msg.addr = src_addr;
+	msg.context = context;
+
+	return fi_ibv_msg_ep_recvmsg(ep, &msg, 0);
+}
+
+static ssize_t
+fi_ibv_msg_ep_sendmsg(struct fid_ep *ep_fid, const struct fi_msg *msg, uint64_t flags)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr = { 0 };
+
+	if (flags & FI_REMOTE_CQ_DATA) {
+		wr.opcode = IBV_WR_SEND_WITH_IMM;
+		wr.imm_data = htonl((uint32_t)msg->data);
+	} else {
+		wr.opcode = IBV_WR_SEND;
+	}
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+	return fi_ibv_send_msg(ep, &wr, msg, flags);
+}
+
+static ssize_t
+fi_ibv_msg_ep_send(struct fid_ep *ep_fid, const void *buf, size_t len,
+		void *desc, fi_addr_t dest_addr, void *context)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr = { 0 };
+
+	wr.opcode = IBV_WR_SEND;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+	wr.send_flags = VERBS_INJECT(ep, len) | VERBS_COMP(ep);
+
+	return fi_ibv_send_buf(ep, &wr, buf, len, desc, context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_senddata(struct fid_ep *ep_fid, const void *buf, size_t len,
+		    void *desc, uint64_t data, fi_addr_t dest_addr, void *context)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr = { 0 };
+
+	wr.opcode = IBV_WR_SEND_WITH_IMM;
+	wr.imm_data = htonl((uint32_t)data);
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+	wr.send_flags = VERBS_INJECT(ep, len) | VERBS_COMP(ep);
+
+	return fi_ibv_send_buf(ep, &wr, buf, len, desc, context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_sendv(struct fid_ep *ep_fid, const struct iovec *iov, void **desc,
+                 size_t count, fi_addr_t dest_addr, void *context)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr = { 0 };
+
+	wr.opcode = IBV_WR_SEND;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+	return fi_ibv_send_iov(ep, &wr, iov, desc, count, context);
+}
+
+static ssize_t fi_ibv_msg_ep_inject(struct fid_ep *ep_fid, const void *buf, size_t len,
+		fi_addr_t dest_addr)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr = { 0 };
+
+	wr.opcode = IBV_WR_SEND;
+	wr.send_flags = IBV_SEND_INLINE;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+
+	return fi_ibv_send_buf_inline(ep, &wr, buf, len);
+}
+
+static ssize_t fi_ibv_msg_ep_injectdata(struct fid_ep *ep_fid, const void *buf, size_t len,
+		    uint64_t data, fi_addr_t dest_addr)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr = { 0 };
+
+	wr.opcode = IBV_WR_SEND_WITH_IMM;
+	wr.imm_data = htonl((uint32_t)data);
+	wr.send_flags = IBV_SEND_INLINE;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+
+	return fi_ibv_send_buf_inline(ep, &wr, buf, len);
+}
+
+struct fi_ops_msg fi_ibv_msg_ep_msg_ops = {
+	.size = sizeof(struct fi_ops_msg),
+	.recv = fi_ibv_msg_ep_recv,
+	.recvv = fi_ibv_msg_ep_recvv,
+	.recvmsg = fi_ibv_msg_ep_recvmsg,
+	.send = fi_ibv_msg_ep_send,
+	.sendv = fi_ibv_msg_ep_sendv,
+	.sendmsg = fi_ibv_msg_ep_sendmsg,
+	.inject = fi_ibv_msg_ep_inject,
+	.senddata = fi_ibv_msg_ep_senddata,
+	.injectdata = fi_ibv_msg_ep_injectdata,
+};
+
+struct fi_ops_msg fi_ibv_msg_srq_ep_msg_ops = {
+	.size = sizeof(struct fi_ops_msg),
+	.recv = fi_no_msg_recv,
+	.recvv = fi_no_msg_recvv,
+	.recvmsg = fi_no_msg_recvmsg,
+	.send = fi_ibv_msg_ep_send,
+	.sendv = fi_ibv_msg_ep_sendv,
+	.sendmsg = fi_ibv_msg_ep_sendmsg,
+	.inject = fi_ibv_msg_ep_inject,
+	.senddata = fi_ibv_msg_ep_senddata,
+	.injectdata = fi_ibv_msg_ep_injectdata,
+};
+
+#define VERBS_COMP_READ_FLAGS(ep, flags) \
+	((!VERBS_SELECTIVE_COMP(ep) || (flags & \
+	  (FI_COMPLETION | FI_TRANSMIT_COMPLETE | FI_DELIVERY_COMPLETE))) ? \
+	   IBV_SEND_SIGNALED : 0)
+#define VERBS_COMP_READ(ep) \
+	VERBS_COMP_READ_FLAGS(ep, ep->info->tx_attr->op_flags)
+
+static ssize_t
+fi_ibv_msg_ep_rma_write(struct fid_ep *ep_fid, const void *buf, size_t len,
+		     void *desc, fi_addr_t dest_addr,
+		     uint64_t addr, uint64_t key, void *context)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+
+	memset(&wr, 0, sizeof(wr));
+	wr.opcode = IBV_WR_RDMA_WRITE;
+	wr.wr.rdma.remote_addr = addr;
+	wr.wr.rdma.rkey = (uint32_t) key;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+	wr.send_flags = VERBS_INJECT(ep, len) | VERBS_COMP(ep);
+
+	return fi_ibv_send_buf(ep, &wr, buf, len, desc, context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_rma_writev(struct fid_ep *ep_fid, const struct iovec *iov, void **desc,
+		      size_t count, fi_addr_t dest_addr,
+		      uint64_t addr, uint64_t key, void *context)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+
+
+	memset(&wr, 0, sizeof(wr));
+	wr.opcode = IBV_WR_RDMA_WRITE;
+	wr.wr.rdma.remote_addr = addr;
+	wr.wr.rdma.rkey = (uint32_t) key;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+	return fi_ibv_send_iov(ep, &wr, iov, desc, count, context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_rma_writemsg(struct fid_ep *ep_fid, const struct fi_msg_rma *msg,
+			uint64_t flags)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+
+	memset(&wr, 0, sizeof(wr));
+	if (flags & FI_REMOTE_CQ_DATA) {
+		wr.opcode = IBV_WR_RDMA_WRITE_WITH_IMM;
+		wr.imm_data = htonl((uint32_t)msg->data);
+	} else {
+		wr.opcode = IBV_WR_RDMA_WRITE;
+	}
+
+	wr.wr.rdma.remote_addr = msg->rma_iov->addr;
+	wr.wr.rdma.rkey = (uint32_t) msg->rma_iov->key;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+	return fi_ibv_send_msg(ep, &wr, msg, flags);
+}
+
+static ssize_t
+fi_ibv_msg_ep_rma_read(struct fid_ep *ep_fid, void *buf, size_t len,
+		    void *desc, fi_addr_t src_addr,
+		    uint64_t addr, uint64_t key, void *context)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+
+	memset(&wr, 0, sizeof(wr));
+	wr.opcode = IBV_WR_RDMA_READ;
+	wr.wr.rdma.remote_addr = addr;
+	wr.wr.rdma.rkey = (uint32_t) key;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+	wr.send_flags = VERBS_COMP_READ(ep);
+
+	return fi_ibv_send_buf(ep, &wr, buf, len, desc, context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_rma_readv(struct fid_ep *ep_fid, const struct iovec *iov, void **desc,
+		     size_t count, fi_addr_t src_addr,
+		     uint64_t addr, uint64_t key, void *context)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr = { 0 };
+	size_t len = 0;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+
+	wr.opcode = IBV_WR_RDMA_READ;
+	wr.wr.rdma.remote_addr = addr;
+	wr.wr.rdma.rkey = (uint32_t) key;
+	wr.send_flags = VERBS_COMP_READ(ep);
+	wr.num_sge = count;
+
+	fi_ibv_set_sge_iov(wr.sg_list, iov, count, desc, len);
+
+	return fi_ibv_send(ep, &wr, context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_rma_readmsg(struct fid_ep *ep_fid, const struct fi_msg_rma *msg,
+			uint64_t flags)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr = { 0 };
+	size_t len = 0;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+
+	wr.opcode = IBV_WR_RDMA_READ;
+	wr.wr.rdma.remote_addr = msg->rma_iov->addr;
+	wr.wr.rdma.rkey = (uint32_t) msg->rma_iov->key;
+	wr.send_flags = VERBS_COMP_READ_FLAGS(ep, flags);
+	wr.num_sge = msg->iov_count;
+
+	fi_ibv_set_sge_iov(wr.sg_list, msg->msg_iov, msg->iov_count, msg->desc, len);
+
+	return fi_ibv_send(ep, &wr, msg->context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_rma_writedata(struct fid_ep *ep_fid, const void *buf, size_t len,
+			void *desc, uint64_t data, fi_addr_t dest_addr,
+			uint64_t addr, uint64_t key, void *context)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+
+	memset(&wr, 0, sizeof(wr));
+	wr.opcode = IBV_WR_RDMA_WRITE_WITH_IMM;
+	wr.imm_data = htonl((uint32_t)data);
+	wr.wr.rdma.remote_addr = addr;
+	wr.wr.rdma.rkey = (uint32_t) key;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+	wr.send_flags = VERBS_INJECT(ep, len) | VERBS_COMP(ep);
+
+	return fi_ibv_send_buf(ep, &wr, buf, len, desc, context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_rma_inject_write(struct fid_ep *ep_fid, const void *buf, size_t len,
+		     fi_addr_t dest_addr, uint64_t addr, uint64_t key)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+
+	memset(&wr, 0, sizeof(wr));
+	wr.opcode = IBV_WR_RDMA_WRITE;
+	wr.wr.rdma.remote_addr = addr;
+	wr.wr.rdma.rkey = (uint32_t) key;
+	wr.send_flags = IBV_SEND_INLINE;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+
+	return fi_ibv_send_buf_inline(ep, &wr, buf, len);
+}
+
+static ssize_t
+fi_ibv_msg_ep_rma_inject_writedata(struct fid_ep *ep_fid, const void *buf, size_t len,
+			uint64_t data, fi_addr_t dest_addr, uint64_t addr,
+			uint64_t key)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+
+	memset(&wr, 0, sizeof(wr));
+	wr.opcode = IBV_WR_RDMA_WRITE_WITH_IMM;
+	wr.imm_data = htonl((uint32_t)data);
+	wr.wr.rdma.remote_addr = addr;
+	wr.wr.rdma.rkey = (uint32_t) key;
+	wr.send_flags = IBV_SEND_INLINE;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+	return fi_ibv_send_buf_inline(ep, &wr, buf, len);
+}
+
+struct fi_ops_rma fi_ibv_msg_ep_rma_ops = {
+	.size = sizeof(struct fi_ops_rma),
+	.read = fi_ibv_msg_ep_rma_read,
+	.readv = fi_ibv_msg_ep_rma_readv,
+	.readmsg = fi_ibv_msg_ep_rma_readmsg,
+	.write = fi_ibv_msg_ep_rma_write,
+	.writev = fi_ibv_msg_ep_rma_writev,
+	.writemsg = fi_ibv_msg_ep_rma_writemsg,
+	.inject = fi_ibv_msg_ep_rma_inject_write,
+	.writedata = fi_ibv_msg_ep_rma_writedata,
+	.injectdata = fi_ibv_msg_ep_rma_inject_writedata,
+};
+
+#define fi_ibv_atomicvalid(name, flags)					\
+static int fi_ibv_msg_ep_atomic_ ## name(struct fid_ep *ep_fid,		\
+			      enum fi_datatype datatype,  		\
+			      enum fi_op op, size_t *count)             \
+{                                                                       \
+	struct fi_ibv_msg_ep *ep = container_of(ep_fid,			\
+						struct fi_ibv_msg_ep,   \
+						ep_fid);                \
+	struct fi_atomic_attr attr;                                     \
+	int ret;                                                        \
+                                                                        \
+	ret = fi_ibv_query_atomic(&ep->domain->util_domain.domain_fid,	\
+				  datatype, op, &attr, flags);		\
+	if (!ret)                                                       \
+		*count = attr.count;                                    \
+	return ret;                                                     \
+}                                                                       \
+
+fi_ibv_atomicvalid(writevalid, 0);
+fi_ibv_atomicvalid(readwritevalid, FI_FETCH_ATOMIC);
+fi_ibv_atomicvalid(compwritevalid, FI_COMPARE_ATOMIC);
+
+int fi_ibv_query_atomic(struct fid_domain *domain_fid, enum fi_datatype datatype,
+			enum fi_op op, struct fi_atomic_attr *attr,
+			uint64_t flags)
+{
+	struct fi_ibv_domain *domain = container_of(domain_fid,
+						    struct fi_ibv_domain,
+						    util_domain.domain_fid);
+	char *log_str_fetch = "fi_fetch_atomic with FI_SUM op";
+	char *log_str_comp = "fi_compare_atomic";
+	char *log_str;
+
+	if (flags & FI_TAGGED)
+		return -FI_ENOSYS;
+
+	if ((flags & FI_FETCH_ATOMIC) && (flags & FI_COMPARE_ATOMIC))
+		return -FI_EBADFLAGS;
+
+	if (!flags) {
+		switch (op) {
+		case FI_ATOMIC_WRITE:
+			break;
+		default:
+			return -FI_ENOSYS;
+		}
+	} else {
+		if (flags & FI_FETCH_ATOMIC) {
+			switch (op) {
+			case FI_ATOMIC_READ:
+				goto check_datatype;
+			case FI_SUM:
+				log_str = log_str_fetch;
+				break;
+			default:
+				return -FI_ENOSYS;
+			}
+		} else if (flags & FI_COMPARE_ATOMIC) {
+			if (op != FI_CSWAP)
+				return -FI_ENOSYS;
+			log_str = log_str_comp;
+		} else {
+			return  -FI_EBADFLAGS;
+		}
+		if (domain->info->tx_attr->op_flags & FI_INJECT) {
+			VERBS_INFO(FI_LOG_EP_DATA,
+				   "FI_INJECT not supported for %s\n", log_str);
+			return -FI_EINVAL;
+		}
+	}
+check_datatype:
+	switch (datatype) {
+	case FI_INT64:
+	case FI_UINT64:
+#if __BITS_PER_LONG == 64
+	case FI_DOUBLE:
+	case FI_FLOAT:
+#endif
+		break;
+	default:
+		return -FI_EINVAL;
+	}
+
+	attr->size = ofi_datatype_size(datatype);
+	if (attr->size == 0)
+		return -FI_EINVAL;
+
+	attr->count = 1;
+	return 0;
+}
+
+static ssize_t
+fi_ibv_msg_ep_atomic_write(struct fid_ep *ep_fid, const void *buf, size_t count,
+			void *desc, fi_addr_t dest_addr, uint64_t addr, uint64_t key,
+			enum fi_datatype datatype, enum fi_op op, void *context)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+	size_t count_copy;
+	int ret;
+
+	if (count != 1)
+		return -FI_E2BIG;
+
+	count_copy = count;
+
+	ret = fi_ibv_msg_ep_atomic_writevalid(ep_fid, datatype, op, &count_copy);
+	if (ret)
+		return ret;
+
+	memset(&wr, 0, sizeof(wr));
+
+	switch(op) {
+	case FI_ATOMIC_WRITE:
+		wr.opcode = IBV_WR_RDMA_WRITE;
+		wr.wr.rdma.remote_addr = addr;
+		wr.wr.rdma.rkey = (uint32_t) (uintptr_t) key;
+		break;
+	default:
+		return -ENOSYS;
+	}
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+
+	wr.send_flags = VERBS_INJECT(ep, sizeof(uint64_t)) | VERBS_COMP(ep) |
+				IBV_SEND_FENCE;
+
+	return fi_ibv_send_buf(ep, &wr, buf, sizeof(uint64_t), desc, context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_atomic_writev(struct fid_ep *ep,
+                        const struct fi_ioc *iov, void **desc, size_t count,
+                        fi_addr_t dest_addr, uint64_t addr, uint64_t key,
+                        enum fi_datatype datatype, enum fi_op op, void *context)
+{
+	if (iov->count != 1)
+		return -FI_E2BIG;
+
+	return fi_ibv_msg_ep_atomic_write(ep, iov->addr, count, desc[0],
+			dest_addr, addr, key, datatype, op, context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_atomic_writemsg(struct fid_ep *ep_fid,
+                        const struct fi_msg_atomic *msg, uint64_t flags)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+	size_t count_copy;
+	int ret;
+
+	if (msg->iov_count != 1 || msg->msg_iov->count != 1)
+		return -FI_E2BIG;
+
+	count_copy = msg->iov_count;
+
+	ret = fi_ibv_msg_ep_atomic_writevalid(ep_fid, msg->datatype, msg->op,
+			&count_copy);
+	if (ret)
+		return ret;
+
+	memset(&wr, 0, sizeof(wr));
+
+	switch (msg->op) {
+	case FI_ATOMIC_WRITE:
+		if (flags & FI_REMOTE_CQ_DATA) {
+			wr.opcode = IBV_WR_RDMA_WRITE_WITH_IMM;
+			wr.imm_data = htonl((uint32_t)msg->data);
+		} else {
+			wr.opcode = IBV_WR_RDMA_WRITE;
+		}
+		wr.wr.rdma.remote_addr = msg->rma_iov->addr;
+		wr.wr.rdma.rkey = (uint32_t) (uintptr_t) msg->rma_iov->key;
+		break;
+	default:
+		return -ENOSYS;
+	}
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+
+	wr.send_flags = VERBS_INJECT_FLAGS(ep, sizeof(uint64_t), flags) |
+				VERBS_COMP_FLAGS(ep, flags) | IBV_SEND_FENCE;
+
+	return fi_ibv_send_buf(ep, &wr, msg->msg_iov->addr, sizeof(uint64_t),
+			msg->desc[0], msg->context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_atomic_readwrite(struct fid_ep *ep_fid, const void *buf, size_t count,
+			void *desc, void *result, void *result_desc,
+			fi_addr_t dest_addr, uint64_t addr, uint64_t key,
+			enum fi_datatype datatype,
+			enum fi_op op, void *context)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+	size_t count_copy;
+	int ret;
+
+	if (count != 1)
+		return -FI_E2BIG;
+
+	count_copy = count;
+
+	ret = fi_ibv_msg_ep_atomic_readwritevalid(ep_fid, datatype, op,
+			&count_copy);
+	if (ret)
+		return ret;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+	memset(&wr, 0, sizeof(wr));
+
+	switch (op) {
+	case FI_ATOMIC_READ:
+		wr.opcode = IBV_WR_RDMA_READ;
+		wr.wr.rdma.remote_addr = addr;
+		wr.wr.rdma.rkey = (uint32_t) (uintptr_t) key;
+		break;
+	case FI_SUM:
+		wr.opcode = IBV_WR_ATOMIC_FETCH_AND_ADD;
+		wr.wr.atomic.remote_addr = addr;
+		wr.wr.atomic.compare_add = (uintptr_t) buf;
+		wr.wr.atomic.swap = 0;
+		wr.wr.atomic.rkey = (uint32_t) (uintptr_t) key;
+		break;
+	default:
+		return -ENOSYS;
+	}
+
+	wr.send_flags = VERBS_COMP(ep) | IBV_SEND_FENCE;
+
+	return fi_ibv_send_buf(ep, &wr, result, sizeof(uint64_t), result_desc,
+		context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_atomic_readwritev(struct fid_ep *ep, const struct fi_ioc *iov,
+			void **desc, size_t count,
+			struct fi_ioc *resultv, void **result_desc,
+			size_t result_count, fi_addr_t dest_addr, uint64_t addr,
+			uint64_t key, enum fi_datatype datatype,
+			enum fi_op op, void *context)
+{
+	if (iov->count != 1)
+		return -FI_E2BIG;
+
+        return fi_ibv_msg_ep_atomic_readwrite(ep, iov->addr, count,
+			desc[0], resultv->addr, result_desc[0],
+			dest_addr, addr, key, datatype, op, context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_atomic_readwritemsg(struct fid_ep *ep_fid,
+				const struct fi_msg_atomic *msg,
+				struct fi_ioc *resultv, void **result_desc,
+				size_t result_count, uint64_t flags)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+	size_t count_copy;
+	int ret;
+
+	if (msg->iov_count != 1 || msg->msg_iov->count != 1)
+		return -FI_E2BIG;
+
+	count_copy = msg->iov_count;
+
+	ret = fi_ibv_msg_ep_atomic_readwritevalid(ep_fid, msg->datatype, msg->op,
+		       &count_copy);
+	if (ret)
+		return ret;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+
+	memset(&wr, 0, sizeof(wr));
+
+	switch (msg->op) {
+	case FI_ATOMIC_READ:
+		wr.opcode = IBV_WR_RDMA_READ;
+		wr.wr.rdma.remote_addr = msg->rma_iov->addr;
+		wr.wr.rdma.rkey = (uint32_t) (uintptr_t) msg->rma_iov->key;
+		break;
+	case FI_SUM:
+		wr.opcode = IBV_WR_ATOMIC_FETCH_AND_ADD;
+		wr.wr.atomic.remote_addr = msg->rma_iov->addr;
+		wr.wr.atomic.compare_add = (uintptr_t) msg->addr;
+		wr.wr.atomic.swap = 0;
+		wr.wr.atomic.rkey = (uint32_t) (uintptr_t) msg->rma_iov->key;
+		break;
+	default:
+		return -ENOSYS;
+	}
+
+	wr.send_flags = VERBS_COMP_FLAGS(ep, flags) | IBV_SEND_FENCE;
+	if (flags & FI_REMOTE_CQ_DATA)
+		wr.imm_data = htonl((uint32_t) msg->data);
+
+	return fi_ibv_send_buf(ep, &wr, resultv->addr, sizeof(uint64_t),
+			result_desc[0], msg->context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_atomic_compwrite(struct fid_ep *ep_fid, const void *buf, size_t count,
+			void *desc, const void *compare,
+			void *compare_desc, void *result,
+			void *result_desc,
+			fi_addr_t dest_addr, uint64_t addr, uint64_t key,
+			enum fi_datatype datatype,
+			enum fi_op op, void *context)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+	size_t count_copy;
+	int ret;
+
+	if (count != 1)
+		return -FI_E2BIG;
+
+	count_copy = count;
+
+	ret = fi_ibv_msg_ep_atomic_compwritevalid(ep_fid, datatype, op, &count_copy);
+	if (ret)
+		return ret;
+
+	memset(&wr, 0, sizeof(wr));
+
+	wr.opcode = IBV_WR_ATOMIC_CMP_AND_SWP;
+	wr.wr.atomic.remote_addr = addr;
+	wr.wr.atomic.compare_add = (uintptr_t) compare;
+	wr.wr.atomic.swap = (uintptr_t) buf;
+	wr.wr.atomic.rkey = (uint32_t) (uintptr_t) key;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+
+	wr.send_flags = VERBS_COMP(ep) | IBV_SEND_FENCE;
+
+	return fi_ibv_send_buf(ep, &wr, result, sizeof(uint64_t), result_desc, context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_atomic_compwritev(struct fid_ep *ep, const struct fi_ioc *iov,
+				void **desc, size_t count,
+				const struct fi_ioc *comparev,
+				void **compare_desc, size_t compare_count,
+				struct fi_ioc *resultv, void **result_desc,
+				size_t result_count,
+				fi_addr_t dest_addr, uint64_t addr, uint64_t key,
+				enum fi_datatype datatype,
+				enum fi_op op, void *context)
+{
+	if (iov->count != 1)
+		return -FI_E2BIG;
+
+	return fi_ibv_msg_ep_atomic_compwrite(ep, iov->addr, count, desc[0],
+				comparev->addr, compare_desc[0], resultv->addr,
+				result_desc[0], dest_addr, addr, key,
+                        	datatype, op, context);
+}
+
+static ssize_t
+fi_ibv_msg_ep_atomic_compwritemsg(struct fid_ep *ep_fid,
+				const struct fi_msg_atomic *msg,
+				const struct fi_ioc *comparev,
+				void **compare_desc, size_t compare_count,
+				struct fi_ioc *resultv,
+				void **result_desc, size_t result_count,
+				uint64_t flags)
+{
+	struct fi_ibv_msg_ep *ep;
+	struct ibv_send_wr wr;
+	size_t count_copy;
+	int ret;
+
+	if (msg->iov_count != 1 || msg->msg_iov->count != 1)
+		return -FI_E2BIG;
+
+	count_copy = msg->iov_count;
+
+	ret = fi_ibv_msg_ep_atomic_compwritevalid(ep_fid, msg->datatype, msg->op,
+		       &count_copy);
+	if (ret)
+		return ret;
+
+	memset(&wr, 0, sizeof(wr));
+
+	wr.opcode = IBV_WR_ATOMIC_CMP_AND_SWP;
+	wr.wr.atomic.remote_addr = msg->rma_iov->addr;
+	wr.wr.atomic.compare_add = (uintptr_t) comparev->addr;
+	wr.wr.atomic.swap = (uintptr_t) msg->addr;
+	wr.wr.atomic.rkey = (uint32_t) (uintptr_t) msg->rma_iov->key;
+
+	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
+
+	wr.send_flags = VERBS_COMP_FLAGS(ep, flags) | IBV_SEND_FENCE;
+
+	if (flags & FI_REMOTE_CQ_DATA)
+		wr.imm_data = htonl((uint32_t) msg->data);
+
+	return fi_ibv_send_buf(ep, &wr, resultv->addr, sizeof(uint64_t),
+			result_desc[0], msg->context);
+}
+
+struct fi_ops_atomic fi_ibv_msg_ep_atomic_ops = {
+	.size		= sizeof(struct fi_ops_atomic),
+	.write		= fi_ibv_msg_ep_atomic_write,
+	.writev		= fi_ibv_msg_ep_atomic_writev,
+	.writemsg	= fi_ibv_msg_ep_atomic_writemsg,
+	.inject		= fi_no_atomic_inject,
+	.readwrite	= fi_ibv_msg_ep_atomic_readwrite,
+	.readwritev	= fi_ibv_msg_ep_atomic_readwritev,
+	.readwritemsg	= fi_ibv_msg_ep_atomic_readwritemsg,
+	.compwrite	= fi_ibv_msg_ep_atomic_compwrite,
+	.compwritev	= fi_ibv_msg_ep_atomic_compwritev,
+	.compwritemsg	= fi_ibv_msg_ep_atomic_compwritemsg,
+	.writevalid	= fi_ibv_msg_ep_atomic_writevalid,
+	.readwritevalid	= fi_ibv_msg_ep_atomic_readwritevalid,
+	.compwritevalid = fi_ibv_msg_ep_atomic_compwritevalid
+};
+
